# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateOrderItem:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CreateOrderItem'
        type: array
    required:
    - items
    type: object
...
  dto.GetOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/dto.OrderResponse'
        type: array
    type: object
  dto.OrderItemResponse:
    properties:
      price:
        type: number
      product:
        $ref: '#/definitions/dto.Product'
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  dto.OrderResponse:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.OrderItemResponse'
        type: array
      order_date:
        type: string
      status:
        type: string
      total_amount:
        type: number
      user_id:
        type: integer
    type: object
  dto.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      sku:
        type: string
    type: object
...
  pkg.ErrorResponseSwag:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  pkg.SuccessResponseSwag:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API cho hệ thống e-commerce
  termsOfService: http://example.com/terms/
  title: E-commerce API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get orders for authenticated user
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order for the authenticated user
      parameters:
      - description: Order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetOrdersResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve order details for the specified order ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pkg.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get order details by order ID
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"